+++ R/integer64.R
@@ -2,7 +2,7 @@
 # R-Code
 # S3 atomic 64bit integers for R
 # (c) 2011-2024 Jens Oehlschägel
-# (c) 2025 Michael Chirico
+# (c) 2025-2026 Michael Chirico
 # Licence: GPL2
 # Provided 'as is', use at your own risk
 # Created: 2011-12-11
@@ -391,6 +391,22 @@ NULL
 #'   all.equal(as.integer64(1:10), as.double(1:10))
 #'   all.equal(as.integer64(1), as.double(1e300))
 #' @name all.equal.integer64
+NULL
+
+#' Factors
+#'
+#' The function [factor] is used to encode a vector as a factor.
+#' 
+#' @inheritParams base::factor
+#' @param nmax an upper bound on the number of levels.
+#'
+#' @return An object of class "factor" or "ordered".
+#' @seealso [factor][base::factor]
+#' @examples
+#'   x <- as.integer64(c(132724613L, -2143220989L, -1L, NA, 1L))
+#'   factor(x)
+#'   ordered(x)
+#' @name factor
 NULL
 
 methods::setOldClass("integer64")
@@ -792,6 +808,71 @@ as.POSIXct.integer64 = function(x, tz="", origin, ...)
 #' @exportS3Method as.POSIXlt integer64
 as.POSIXlt.integer64 = function(x, tz="", origin, ...)
   as.POSIXlt(as.double(x, ...), tz=tz, origin=origin, ...)
+
+#' @rdname as.character.integer64
+#' @export as.factor
+as.factor = function(x) factor(x=x)
+
+#' @rdname as.character.integer64
+#' @export as.ordered
+as.ordered = function(x) ordered(x=x)
+
+#' @rdname factor
+#' @export
+factor = function(x=character(), levels, labels=levels, exclude=NA, ordered=is.ordered(x), nmax=NA) {
+  if (!is.integer64(x)) {
+    sys_call = sys.call()
+    sys_call[[1L]] = base::factor
+    pf = parent.frame()
+    return(withCallingHandlers_and_choose_call(eval(sys_call, envir=pf), "factor"))
+  }
+  
+  nx = names(x)
+  if (missing(levels)) {
+    levels = sort(unique(x))
+  } else if (length(x) >= 4000) {
+    levels = as.integer64(levels)
+  }
+  # use base::factor for short vectors because it is faster
+  if (length(x) < 4000) {
+    force(ordered)
+    x = as.character(x)
+    levels = as.character(levels)
+    if (missing(labels))
+      return(withCallingHandlers_and_choose_call(base::factor(x=x, levels=levels, exclude=exclude, ordered=ordered, nmax=nmax), "factor"))
+    else
+      return(withCallingHandlers_and_choose_call(base::factor(x=x, levels=levels, labels=labels, exclude=exclude, ordered=ordered, nmax=nmax), "factor"))
+  }
+
+  # basically copied from base::factor, but using the benefit from caching
+  levels = levels[is.na(match(levels, exclude))]
+  ret = match(x, levels)
+  if (!is.null(nx)) 
+    names(ret) = nx
+  if (missing(labels)) {
+    levels(ret) = as.character(levels)
+  } else {
+    nlab = length(labels)
+    if (nlab == length(levels)) {
+      xlevs = as.character(labels)
+      nlevs = unique(xlevs)
+      at = attributes(ret)
+      at$levels = nlevs
+      ret = match(xlevs, nlevs)[ret]
+      attributes(ret) = at
+    } else if (nlab == 1L) {
+      levels(ret) = paste0(labels, seq_along(levels))
+    } else {
+      stop(gettextf("invalid 'labels'; length %d should be 1 or %d", nlab, length(levels), domain="R-base"), domain=NA)
+    }
+  }
+  class(ret) <- c(if (ordered) "ordered", "factor")
+  ret
+}
+
+#' @rdname factor
+#' @export
+ordered = function(x=character(), ...) factor(x, ..., ordered=TRUE)
 
 #' @rdname as.character.integer64
 #' @export